Add deployment script and config for ropsten
