+++ hardhat.config.js
@@ -1,6 +1,9 @@
 require("@nomiclabs/hardhat-waffle");
 require("@nomiclabs/hardhat-web3");
 
+// read mnemonic from bridge config
+require('dotenv').config({ path: 'node_modules/bridge/.env.development' });
+
 // This is a sample Hardhat task. To learn how to create your own go to
 // https://hardhat.org/guides/create-task.html
 task("accounts", "Prints the list of accounts", async () => {
@@ -23,6 +26,15 @@ task("accountsFromWeb3", "Prints accounts", async (_, {web3}) => {
  * @type import('hardhat/config').HardhatUserConfig
  */
 module.exports = {
+    hardhat: {},
+    networks: {
+        ropsten: {
+            accounts: {
+                mnemonic: process.env.REACT_APP_DEV_MNEMONIC,
+            },
+            url: process.env.ROPSTEN_URL,
+        },
+    },
     solidity: {
         compilers: [
             // for legacy code (azimuth)
