+++ package.json
@@ -8,7 +8,10 @@
     "@nomiclabs/hardhat-web3": "^2.0.0",
     "@openzeppelin/test-environment": "^0.1.9",
     "@openzeppelin/test-helpers": "^0.5.11",
+    "bridge": "git+https://github.com/urbit/bridge.git",
     "chai": "^4.3.4",
+    "dotenv": "^10.0.0",
+    "esm": "^3.2.25",
     "ethereum-waffle": "^3.3.0",
     "ethers": "^5.1.4",
     "hardhat": "^2.2.1",
